Rename core.py -> primitives.py
