+++ python/sodiumfrp/__init__.py
@@ -1,4 +1,10 @@
-from .core import Cell, CellLoop, CellSink, Stream, StreamLoop, StreamSink
+from .primitives import \
+    Cell, \
+    CellLoop, \
+    CellSink, \
+    Stream, \
+    StreamLoop, \
+    StreamSink
 from .listener import Listener
 from .transaction import Transaction
 from .unit import Unit, UNIT
